adds defensive channel reconnect after api session change
Previously it was possible for channels to remain connected to Edge
Routers with previous API Sessions. This would occur in scenarios
where the SDK requested a new API Session but did not invalidate the
previous one. Under normal conditions this shouldn't happen, but bugs
have caused it to happen in the past.