Match of cidr /0 fails as a result of left shift undefined behavior.
When the prefix length is 0, the expression -1UL << 32 is evaluated. A
shift of 32 or more is undefined on a 32-bit integer, meaning the
behavior is unpredictable (as not prescribed) and is likely compiler and
hardware dependent. See C17 SS 6.5.7 for clarity.  On amd64, the
result of the expression is -1UL.

Signed-off-by: Tom Carroll <4632752+tomc797@users.noreply.github.com>